Fix: Test fixture directories
