+++ test/Fixture/ClassesHaveTests/NotAllClassesHaveTests/Src/AnotherExampleClass.php
@@ -11,8 +11,8 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\NotAllClassesHaveTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\NotAllClassesHaveTests\Src;
 
-final class OneMoreExampleClass
+final class AnotherExampleClass
 {
 }
