+++ test/Fixture/ClassesHaveTests/NotAllClassesHaveTests/Src/ExampleClass.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\NotAllClassesHaveTests\Src;
 
 final class ExampleClass
 {
