+++ test/Fixture/ClassesHaveTests/NotAllClassesHaveTests/Src/OneMoreExampleClass.php
@@ -11,8 +11,8 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\NotAllClassesHaveTests\Src;
 
-abstract class AbstractClass
+final class OneMoreExampleClass
 {
 }
