+++ test/Fixture/ClassesHaveTests/WithTests/Src/ExampleTrait.php
@@ -11,8 +11,8 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\NotAllClassesHaveTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithTests\Src;
 
-final class ExampleClass
+trait ExampleTrait
 {
 }
