+++ test/Fixture/ClassesHaveTests/WithoutTests/Src/ExampleClass.php
@@ -11,8 +11,8 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithoutTests\Src;
 
-interface ExampleInterface
+final class ExampleClass
 {
 }
