+++ test/Fixture/ClassesHaveTests/WithoutTests/Src/ExampleInterface.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/test-util
  */
 
-namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithoutTests;
+namespace Localheinz\Test\Util\Test\Fixture\ClassesHaveTests\WithoutTests\Src;
 
 interface ExampleInterface
 {
